fix(mobile-styles): fix mobile styles in ValidatorsList.tsx
- Fix import path for RoundedGradientImage component
- Add useIsMobile hook import
- Add styles for different font sizes based on mobile/desktop view
- Update Avatar component to use RoundedGradientImage component
- Update font styles for moniker text
- Update font styles for votingPower text
- Update thousandSeparator function to numFormatter function
