+++ packages/hooks/useValidators.ts
@@ -146,7 +146,6 @@ const sortByVotingPower = (a: ValidatorInfo, b: ValidatorInfo) =>
 const formatValidators = (vals: ValidatorInfo[]) => {
   let i = 0;
   for (const v of vals) {
-    v.votingPower += " TORI";
     v.rank = (i + 1).toString();
     i++;
   }
