+++ packages/screens/Marketplace/MarketplaceScreen.tsx
@@ -38,6 +38,7 @@ import {
   fontSemibold28,
 } from "../../utils/style/fonts";
 import { layout, screenContentMaxWidthLarge } from "../../utils/style/layout";
+import { numFormatter } from "../../utils/text";
 import { arrayIncludes } from "../../utils/typescript";
 
 const TABLE_ROWS = {
@@ -392,34 +393,6 @@ const InnerCell: React.FC<{
   );
 };
 
-function nFormatter(
-  num: number | undefined | string,
-  digits: number | undefined
-) {
-  const lookup = [
-    { value: 1, symbol: "" },
-    { value: 1e3, symbol: "K" },
-    { value: 1e6, symbol: "M" },
-    { value: 1e9, symbol: "G" },
-    { value: 1e12, symbol: "T" },
-    { value: 1e15, symbol: "P" },
-    { value: 1e18, symbol: "E" },
-  ];
-  const rx = /\.0+$|(\.[0-9]*[1-9])0+$/;
-  const item = lookup
-    .slice()
-    .reverse()
-    .find(function (item) {
-      // @ts-expect-error
-      return num >= item.value;
-    });
-
-  return item
-    ? // @ts-expect-error
-      (num / item.value).toFixed(digits).replace(rx, "$1") + item.symbol
-    : "0";
-}
-
 interface RowData {
   id: string;
   rank: number;
@@ -457,17 +430,17 @@ const useRowData = (collection: Collection, rank: number): RowData => {
       denom: collection.denom,
     },
     TimePeriodPercentualVolume: getDelta(collection),
-    sales: nFormatter(collection.numTrades, 0),
+    sales: numFormatter(collection.numTrades, 0),
     floorPrice: {
       networkId: collection.networkId,
       value: collection.floorPrice,
       denom: collection.denom,
     },
-    owners: nFormatter(
+    owners: numFormatter(
       collection.numOwners,
       collection.numOwners.toString().length
     ),
-    supply: nFormatter(
+    supply: numFormatter(
       collection.maxSupply,
       collection.maxSupply.toString().length
     ),
