+++ packages/utils/text.tsx
@@ -65,3 +65,31 @@ export const replaceTextWithComponent = (
 
 export const pluralOrNot = (word: string, quantity: number) =>
   quantity > 1 ? word + "s" : word;
+
+export function numFormatter(
+  num: number | undefined | string,
+  digits: number | undefined
+) {
+  const lookup = [
+    { value: 1, symbol: "" },
+    { value: 1e3, symbol: "K" },
+    { value: 1e6, symbol: "M" },
+    { value: 1e9, symbol: "G" },
+    { value: 1e12, symbol: "T" },
+    { value: 1e15, symbol: "P" },
+    { value: 1e18, symbol: "E" },
+  ];
+  const rx = /\.0+$|(\.[0-9]*[1-9])0+$/;
+  const item = lookup
+    .slice()
+    .reverse()
+    .find(function (item) {
+      // @ts-expect-error
+      return num >= item.value;
+    });
+
+  return item
+    ? // @ts-expect-error
+      (num / item.value).toFixed(digits).replace(rx, "$1") + item.symbol
+    : "0";
+}
